Mejoradas las notificaciones y otras cosas más

+++ backend/controllers/archivos.js
@@ -467,13 +467,10 @@ const borrarArchivoBBDD = async (req, res) => {
     // Borro el archivo
     await Archivo.deleteMany({ id: idArchivo });
 
-    // Si no he encontrado otros archivos con ese padre, borro el padre
+    // Si no he encontrado otros archivos con ese padre, 
+    // borro el padre ( siempre que no sea la carpeta "nuevas")
     var ficherosCarpeta = await Archivo.find({ parents: idParentFile });
-    console.log("ficherosCarpeta", ficherosCarpeta);
-    console.log("ficherosCarpeta.length", ficherosCarpeta.length);
-
-    if (ficherosCarpeta.length == 0) {
-      console.log("No hay otros archivos en esa carpeta, así que la borro");
+    if (ficherosCarpeta.length == 0 && (idParentFile !== '1y0wM99USkqzvObuf8HY43kOw6tAbJQ0s') ) {
       await Etiqueta.deleteMany({ id: idParentFile });
       try {
         return drive.files
